+++ packages/backend/src/libp2p/connectionsManager.ts
@@ -116,7 +116,7 @@ export interface InitLibp2pParams {
   certs: Certificates
 }
 
-export enum IsTorInit{
+export enum TorInitState{
   STARTING = 'starting',
   STARTED = 'started',
   NOT_STARTED = 'not-started'
@@ -143,7 +143,7 @@ export class ConnectionsManager extends EventEmitter {
   localStorage: LocalDB
   communityState: ServiceState
   registrarState: ServiceState
-  isTorInit: IsTorInit = IsTorInit.NOT_STARTED
+  isTorInit: TorInitState = TorInitState.NOT_STARTED
 
   constructor({ options, socketIOPort, httpTunnelPort, torControlPort, torAuthCookie, torResourcesPath, torBinaryPath }: IConstructor) {
     super()
@@ -230,12 +230,16 @@ export class ConnectionsManager extends EventEmitter {
 
     await this.dataServer.listen()
 
+    if (this.torControlPort) {
+      await this.launchCommunityFromStorage()
+    }
+
     this.io.on('connection', async() => {
-      if (this.isTorInit === IsTorInit.STARTED || this.isTorInit === IsTorInit.STARTING) return
-      this.isTorInit = IsTorInit.STARTING
+      if (this.isTorInit === TorInitState.STARTED || this.isTorInit === TorInitState.STARTING) return
+      this.isTorInit = TorInitState.STARTING
         if (this.torBinaryPath) {
           await this.tor.init()
-          this.isTorInit = IsTorInit.STARTED
+          this.isTorInit = TorInitState.STARTED
         }
         await this.launchCommunityFromStorage()
     })
@@ -323,7 +327,6 @@ export class ConnectionsManager extends EventEmitter {
 
     if (this.torControlPort) {
       this.tor.initTorControl()
-      await this.launchCommunityFromStorage()
     } else if (this.torBinaryPath) {
       // Tor init will be executed on connection event
     } else {
@@ -561,6 +564,10 @@ export class ConnectionsManager extends EventEmitter {
     this.dataServer.on(
       SocketActionTypes.REGISTER_USER_CERTIFICATE,
       async (args: RegisterUserCertificatePayload) => {
+        if (!this.socksProxyAgent) {
+          this.socksProxyAgent = this.createAgent()
+        }
+
         await this.registration.sendCertificateRegistrationRequest(
           args.serviceAddress,
           args.userCsr,