+++ packages/desktop/src/renderer/components/LoadingPanel/LoadingPanel.tsx
@@ -35,7 +35,6 @@ const LoadingPanel = () => {
 
   const community = useSelector(communities.selectors.currentCommunity)
   const owner = Boolean(community?.CA)
-
   const currentIdentity = useSelector(identity.selectors.currentIdentity)
   const usersData = Object.keys(useSelector(users.selectors.certificates))
   const isOnlyOneUser = usersData.length === 1
@@ -95,6 +94,7 @@ const LoadingPanel = () => {
         {...loadingPanelModal}
         openUrl={openUrl}
         torConnectionInfo={torConnectionProcessSelector}
+        isOwner={owner}
       />
     )
   }