+++ packages/desktop/src/renderer/components/Settings/Tabs/QRCode/QRCode.tsx
@@ -3,11 +3,9 @@ import { useSelector } from 'react-redux'
 import { communities } from '@quiet/state-manager'
 
 import { QRCodeComponent } from './QRCode.component'
+import { Site } from '@quiet/common'
 
 export const QRCode: React.FC = () => {
-  const community = useSelector(communities.selectors.currentCommunity)
-  const invitationLink =
-    `https://tryquiet.org/join?code=${community?.registrarUrl}` || 'https://tryquiet.org/'
-
+  const invitationLink = useSelector(communities.selectors.invitationUrl) || Site.MAIN_PAGE
   return <QRCodeComponent value={invitationLink} />
 }