+++ packages/e2e-tests/leaveCommunity.crossplatform.test.ts
@@ -11,16 +11,20 @@ import {
 } from './selectors.crossplatform'
 import logger from './logger'
 import { BuildSetup } from './crossplatform.utils'
-import getPort from 'get-port'
-const log = logger('newUser:')
+
+const log = logger('leaveCommunity:')
 
 jest.setTimeout(900000)
-describe('New User', () => {
+describe('Leave Community', () => {
   let buildSetup: BuildSetup
   let driver: ThenableWebDriver
+  const port = 9520
+  const debugPort = 9521
 
   let buildSetup2: BuildSetup
   let driver2: ThenableWebDriver
+  const port2 = 9522
+  const debugPort2 = 9523
 
   let generalChannel: Channel
   let generalChannel2: Channel
@@ -30,11 +34,13 @@ describe('New User', () => {
   const communityName = 'testcommunity'
   const ownerUsername = 'bob'
   const ownerMessages = ['Hi']
-  const joiningUserUsername = 'alice-joining'
+
   const joiningUserMessages = ['Nice to meet you all']
+
+  const joiningUserUsername1 = 'alice1'
+  const joiningUserUsername2 = 'alice2'
+
   beforeAll(async () => {
-    const port = await getPort()
-    const debugPort = await getPort()
     buildSetup = new BuildSetup({ port, debugPort })
     await buildSetup.createChromeDriver()
     driver = buildSetup.getDriver()
@@ -93,6 +99,7 @@ describe('New User', () => {
       expect(isGeneralChannel).toBeTruthy()
       expect(generalChannelText).toEqual('# general')
     })
+
     it('Send message', async () => {
       const isMessageInput = await generalChannel.messageInput.isDisplayed()
       expect(isMessageInput).toBeTruthy()
@@ -110,14 +117,11 @@ describe('New User', () => {
       await settingsModal.switchTab('invite') // TODO: Fix - the invite tab should be default for the owner
       const invitationCodeElement = await settingsModal.invitationCode()
       invitationCode = await invitationCodeElement.getText()
-      console.log({ invitationCode })
       log('Received invitation code:', invitationCode)
       await settingsModal.close()
     })
     it('Guest setup', async () => {
       console.log('Second client')
-      const port2 = await getPort()
-      const debugPort2 = await getPort()
       buildSetup2 = new BuildSetup({ port: port2, debugPort: debugPort2 })
       await buildSetup2.createChromeDriver()
       driver2 = buildSetup2.getDriver()
@@ -130,47 +134,74 @@ describe('New User', () => {
       })
     }
     it('StartingLoadingPanel modal', async () => {
-      console.log('new user - 2')
       const loadingPanel = new StartingLoadingPanel(driver2)
       const isLoadingPanel = await loadingPanel.element.isDisplayed()
       expect(isLoadingPanel).toBeTruthy()
     })
     it('Guest joins the new community successfully', async () => {
-      console.log('new user - 3')
       const joinCommunityModal = new JoinCommunityModal(driver2)
       const isJoinCommunityModal = await joinCommunityModal.element.isDisplayed()
       expect(isJoinCommunityModal).toBeTruthy()
-      console.log({ invitationCode })
       await joinCommunityModal.typeCommunityCode(invitationCode)
       await joinCommunityModal.submit()
     })
     it('RegisterUsernameModal', async () => {
-      console.log('new user - 4')
       const registerModal2 = new RegisterUsernameModal(driver2)
       const isRegisterModal2 = await registerModal2.element.isDisplayed()
       expect(isRegisterModal2).toBeTruthy()
-      await registerModal2.typeUsername(joiningUserUsername)
+      await registerModal2.typeUsername(joiningUserUsername1)
       await registerModal2.submit()
     })
     it.skip('JoiningLoadingPanel', async () => {
-      console.log('new user - 5')
       const loadingPanelCommunity2 = new JoiningLoadingPanel(driver)
       const isLoadingPanelCommunity2 = await loadingPanelCommunity2.element.isDisplayed()
       expect(isLoadingPanelCommunity2).toBeTruthy()
     })
-    it('User sends a message', async () => {
-      console.log('new user - 6')
+    it('Channels are visible', async () => {
       generalChannel2 = new Channel(driver2, 'general')
       await generalChannel2.element.isDisplayed()
       const isMessageInput2 = await generalChannel2.messageInput.isDisplayed()
       expect(isMessageInput2).toBeTruthy()
-      console.log('FETCHING CHANNEL MESSAGES!')
-      await new Promise<void>(resolve => setTimeout(() => resolve(), 15000))
+
+      await new Promise<void>(resolve => setTimeout(() => resolve(), 5000))
+    })
+    it('Leave community', async () => {
+      const settingsModal = await new Sidebar(driver2).openSettings()
+      const isSettingsModal = await settingsModal.element.isDisplayed()
+      expect(isSettingsModal).toBeTruthy()
+      await settingsModal.openLeaveCommunityModal()
+      await settingsModal.leaveCommunityButton()
+    })
+    if (process.env.TEST_MODE) {
+      it('Close debug modal', async () => {
+        const debugModal = new DebugModeModal(driver2)
+        await debugModal.close()
+      })
+    }
+    it('Guest re-join to community successfully', async () => {
+      const joinCommunityModal = new JoinCommunityModal(driver2)
+      const isJoinCommunityModal = await joinCommunityModal.element.isDisplayed()
+      expect(isJoinCommunityModal).toBeTruthy()
+      await joinCommunityModal.typeCommunityCode(invitationCode)
+      await joinCommunityModal.submit()
+    })
+    it('Guest register new username', async () => {
+      const registerModal2 = new RegisterUsernameModal(driver2)
+      const isRegisterModal2 = await registerModal2.element.isDisplayed()
+      expect(isRegisterModal2).toBeTruthy()
+      await registerModal2.typeUsername(joiningUserUsername2)
+      await registerModal2.submit()
+    })
+    it('Guest sends a message', async () => {
+      generalChannel2 = new Channel(driver2, 'general')
+      await generalChannel2.element.isDisplayed()
+      const isMessageInput2 = await generalChannel2.messageInput.isDisplayed()
+      expect(isMessageInput2).toBeTruthy()
+      await new Promise<void>(resolve => setTimeout(() => resolve(), 5000))
       await generalChannel2.sendMessage(joiningUserMessages[0])
     })
     it('Sent message is visible in a channel', async () => {
-      console.log('new user - 7')
-      const messages2 = await generalChannel2.getUserMessages(joiningUserUsername)
+      const messages2 = await generalChannel2.getUserMessages(joiningUserUsername2)
       const text2 = await messages2[0].getText()
       expect(text2).toEqual(joiningUserMessages[0])
     })