+++ packages/e2e-tests/selectors.crossplatform.ts
@@ -18,6 +18,26 @@ export class StartingLoadingPanel {
   //   return this.driver.findElement(By.xpath(`//span[text()="${this.text}"]`))
   // }
 }
+
+export class WarningModal {
+  private readonly driver: ThenableWebDriver
+
+  constructor(driver: ThenableWebDriver) {
+    this.driver = driver
+  }
+
+  get titleElement() {
+    return this.driver.wait(
+      until.elementLocated(By.xpath('//h3[@data-testid="warningModalTitle"]'))
+    )
+  }
+
+  async close() {
+    const submitButton = await this.driver.findElement(By.xpath('//button[@data-testid="warningModalSubmit"]'))
+    await submitButton.click()
+  }
+}
+
 export class JoiningLoadingPanel {
   private readonly driver: ThenableWebDriver
   constructor(driver: ThenableWebDriver) {
@@ -201,6 +221,20 @@ export class Settings {
     return this.driver.wait(until.elementLocated(By.xpath("//h6[text()='Settings']")))
   }
 
+  async openLeaveCommunityModal() {
+    const tab = await this.driver.wait(
+      until.elementLocated(By.xpath('//p[@data-testid="leave-community-tab"]'))
+    )
+    await tab.click()
+  }
+
+  async leaveCommunityButton() {
+    const button = await this.driver.wait(
+      until.elementLocated(By.xpath('//button[text()="Leave community"]'))
+    )
+    await button.click()
+  }
+
   async switchTab(name: string) {
     const tab = await this.driver.findElement(
       By.xpath(`//button[@data-testid='${name}-settings-tab']`)
@@ -229,6 +263,7 @@ export class DebugModeModal {
   private readonly driver: ThenableWebDriver
   constructor(driver: ThenableWebDriver) {
     this.driver = driver
+    console.log('Debug modal')
   }
 
   get element() {
@@ -239,5 +274,22 @@ export class DebugModeModal {
 
   get button() {
     return this.driver.wait(until.elementLocated(By.xpath("//button[text()='Understand']")))
+  }
+
+  async close() {
+    console.log('Closing debug modal')
+    await this.element.isDisplayed()
+    const button = await this.button
+    console.log('Debug modal title is displayed')
+    await button.isDisplayed()
+    console.log('Button is displayed')
+    await button.click()
+    console.log('Button click')
+    try {
+      const log = await this.driver.executeScript('arguments[0].click();', button)
+      console.log('executeScript', log)
+    } catch (e) {
+      console.log('Probably click properly close modal')
+    }
   }
 }