+++ packages/backend/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.2.0-alpha.22](https://github.com/TryQuiet/backend/compare/@quiet/backend@1.2.0-alpha.21...@quiet/backend@1.2.0-alpha.22) (2023-04-12)
+
+**Note:** Version bump only for package @quiet/backend
+
+
+
+
+
+# [1.2.0-alpha.21](https://github.com/TryQuiet/backend/compare/@quiet/backend@1.2.0-alpha.20...@quiet/backend@1.2.0-alpha.21) (2023-04-07)
+
+**Note:** Version bump only for package @quiet/backend
+
+
+
+
+
+# [1.2.0-alpha.20](https://github.com/TryQuiet/backend/compare/@quiet/backend@1.2.0-alpha.14...@quiet/backend@1.2.0-alpha.20) (2023-04-07)
+
+**Note:** Version bump only for package @quiet/backend
+
+
+
+
+
 # [1.2.0-alpha.14](https://github.com/TryQuiet/backend/compare/@quiet/backend@1.2.0-alpha.0...@quiet/backend@1.2.0-alpha.14) (2023-03-30)
 
 **Note:** Version bump only for package @quiet/backend