+++ packages/desktop/CHANGELOG.md
@@ -3,6 +3,54 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.2.0-alpha.25](https://github.com/TryQuiet/quiet/compare/quiet@1.2.0-alpha.24...quiet@1.2.0-alpha.25) (2023-04-12)
+
+**Note:** Version bump only for package quiet
+
+
+
+
+
+# [1.2.0-alpha.24](https://github.com/ZbayApp/monorepo/compare/quiet@1.2.0-alpha.23...quiet@1.2.0-alpha.24) (2023-04-07)
+
+**Note:** Version bump only for package quiet
+
+
+
+
+
+# [1.2.0-alpha.23](https://github.com/TryQuiet/quiet/compare/quiet@1.2.0-alpha.22...quiet@1.2.0-alpha.23) (2023-04-07)
+
+**Note:** Version bump only for package quiet
+
+
+
+
+
+# [1.2.0-alpha.22](https://github.com/TryQuiet/quiet/compare/quiet@1.2.0-alpha.21...quiet@1.2.0-alpha.22) (2023-04-07)
+
+**Note:** Version bump only for package quiet
+
+
+
+
+
+# [1.2.0-alpha.21](https://github.com/TryQuiet/quiet/compare/quiet@1.2.0-alpha.20...quiet@1.2.0-alpha.21) (2023-04-07)
+
+**Note:** Version bump only for package quiet
+
+
+
+
+
+# [1.2.0-alpha.20](https://github.com/TryQuiet/quiet/compare/quiet@1.2.0-alpha.19...quiet@1.2.0-alpha.20) (2023-04-07)
+
+**Note:** Version bump only for package quiet
+
+
+
+
+
 # [1.2.0-alpha.19](/compare/quiet@1.2.0-alpha.18...quiet@1.2.0-alpha.19) (2023-03-30)
 
 **Note:** Version bump only for package quiet