+++ packages/e2e-tests/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.2.0-alpha.21](https://github.com/TryQuiet/quiet/compare/e2e-tests@1.2.0-alpha.20...e2e-tests@1.2.0-alpha.21) (2023-04-12)
+
+**Note:** Version bump only for package e2e-tests
+
+
+
+
+
+# [1.2.0-alpha.20](https://github.com/TryQuiet/quiet/compare/e2e-tests@1.2.0-alpha.14...e2e-tests@1.2.0-alpha.20) (2023-04-07)
+
+**Note:** Version bump only for package e2e-tests
+
+
+
+
+
 # [1.2.0-alpha.14](https://github.com/TryQuiet/quiet/compare/e2e-tests@1.2.0-alpha.0...e2e-tests@1.2.0-alpha.14) (2023-03-30)
 
 **Note:** Version bump only for package e2e-tests